repo reorg: move HCM filter and various cleanups
The core connection manager code has been left in the Http
namespace since it is required for admin. Most of this change
is various cleanups when I realized that the Http mocks were
pulling in the full conn_manager_impl dependency tree. This is
now fixed so that the Http mocks are based on pure virtual
interfaces now.
