+++ source/extensions/extensions_build_config.bzl
@@ -35,6 +35,7 @@ EXTENSIONS = {
     "envoy.filters.network.client_ssl_auth":    "//source/extensions/filters/network/client_ssl_auth:config",
     "envoy.filters.network.echo":               "//source/extensions/filters/network/echo:config",
     "envoy.filters.network.ext_authz":          "//source/extensions/filters/network/ext_authz:config",
+    "envoy.filters.network.hcm":                "//source/extensions/filters/network/hcm:config",
     "envoy.filters.network.mongo_proxy":        "//source/extensions/filters/network/mongo_proxy:config",
     "envoy.filters.network.redis_proxy":        "//source/extensions/filters/network/redis_proxy:config",
     "envoy.filters.network.ratelimit":          "//source/extensions/filters/network/ratelimit:config",
