+++ source/server/http/BUILD
@@ -40,6 +40,7 @@ envoy_cc_library(
         "//source/common/http:codes_lib",
         "//source/common/http:conn_manager_lib",
         "//source/common/http:date_provider_lib",
+        "//source/common/http:default_server_string_lib",
         "//source/common/http:header_map_lib",
         "//source/common/http:headers_lib",
         "//source/common/http:utility_lib",
@@ -50,7 +51,6 @@ envoy_cc_library(
         "//source/common/router:config_lib",
         "//source/common/upstream:host_utility_lib",
         "//source/extensions/access_loggers/file:file_access_log_lib",
-        "//source/server/config/network:http_connection_manager_lib",
         "@envoy_api//envoy/admin/v2:config_dump_cc",
     ],
 )
@@ -76,6 +76,7 @@ envoy_cc_library(
         "//include/envoy/http:codes_interface",
         "//include/envoy/http:filter_interface",
         "//include/envoy/http:header_map_interface",
+        "//include/envoy/server:filter_config_interface",
         "//source/common/common:assert_lib",
         "//source/common/common:enum_to_int",
         "//source/common/config:filter_json_lib",
@@ -85,7 +86,6 @@ envoy_cc_library(
         "//source/common/http:headers_lib",
         "//source/common/http:utility_lib",
         "//source/common/protobuf:utility_lib",
-        "//source/server/config/network:http_connection_manager_lib",
         "@envoy_api//envoy/config/filter/http/health_check/v2:health_check_cc",
     ],
 )
