+++ source/server/http/health_check.cc
@@ -17,8 +17,6 @@
 #include "common/http/utility.h"
 #include "common/protobuf/utility.h"
 
-#include "server/config/network/http_connection_manager.h"
-
 namespace Envoy {
 namespace Server {
 namespace Configuration {
