+++ test/common/http/conn_manager_utility_test.cc
@@ -27,6 +27,41 @@ using testing::_;
 namespace Envoy {
 namespace Http {
 
+class MockConnectionManagerConfig : public ConnectionManagerConfig {
+public:
+  MockConnectionManagerConfig() { ON_CALL(*this, generateRequestId()).WillByDefault(Return(true)); }
+
+  // Http::ConnectionManagerConfig
+  ServerConnectionPtr createCodec(Network::Connection& connection, const Buffer::Instance& instance,
+                                  ServerConnectionCallbacks& callbacks) override {
+    return ServerConnectionPtr{createCodec_(connection, instance, callbacks)};
+  }
+
+  MOCK_METHOD0(accessLogs, const std::list<AccessLog::InstanceSharedPtr>&());
+  MOCK_METHOD3(createCodec_, ServerConnection*(Network::Connection&, const Buffer::Instance&,
+                                               ServerConnectionCallbacks&));
+  MOCK_METHOD0(dateProvider, DateProvider&());
+  MOCK_METHOD0(drainTimeout, std::chrono::milliseconds());
+  MOCK_METHOD0(filterFactory, FilterChainFactory&());
+  MOCK_METHOD0(generateRequestId, bool());
+  MOCK_METHOD0(idleTimeout, const absl::optional<std::chrono::milliseconds>&());
+  MOCK_METHOD0(routeConfigProvider, Router::RouteConfigProvider&());
+  MOCK_METHOD0(serverName, const std::string&());
+  MOCK_METHOD0(stats, ConnectionManagerStats&());
+  MOCK_METHOD0(tracingStats, ConnectionManagerTracingStats&());
+  MOCK_METHOD0(useRemoteAddress, bool());
+  MOCK_CONST_METHOD0(xffNumTrustedHops, uint32_t());
+  MOCK_METHOD0(forwardClientCert, Http::ForwardClientCertType());
+  MOCK_CONST_METHOD0(setCurrentClientCertDetails,
+                     const std::vector<Http::ClientCertDetailsType>&());
+  MOCK_METHOD0(localAddress, const Network::Address::Instance&());
+  MOCK_METHOD0(userAgent, const absl::optional<std::string>&());
+  MOCK_METHOD0(tracingConfig, const Http::TracingConnectionManagerConfig*());
+  MOCK_METHOD0(listenerStats, ConnectionManagerListenerStats&());
+  MOCK_CONST_METHOD0(proxy100Continue, bool());
+  MOCK_CONST_METHOD0(http1Settings, const Http::Http1Settings&());
+};
+
 class ConnectionManagerUtilityTest : public testing::Test {
 public:
   ConnectionManagerUtilityTest() {
