+++ test/common/network/BUILD
@@ -84,6 +84,7 @@ envoy_cc_test(
         "//source/common/upstream:upstream_includes",
         "//source/common/upstream:upstream_lib",
         "//source/extensions/filters/network/ratelimit:ratelimit_lib",
+        "//source/extensions/filters/network/tcp_proxy",
         "//test/common/upstream:utility_lib",
         "//test/mocks/buffer:buffer_mocks",
         "//test/mocks/network:network_mocks",
