+++ test/common/router/router_upstream_log_test.cc
@@ -2,6 +2,7 @@
 #include <regex>
 
 #include "common/config/filter_json.h"
+#include "common/network/utility.h"
 #include "common/router/router.h"
 #include "common/upstream/upstream_impl.h"
