+++ test/integration/BUILD
@@ -217,6 +217,7 @@ envoy_cc_test_library(
     ],
     deps = [
         ":integration_lib",
+        "//source/extensions/filters/network/hcm:config",
         "//test/common/upstream:utility_lib",
     ],
 )
@@ -302,7 +303,6 @@ envoy_cc_test_library(
         "//source/server/config/http:gzip_lib",
         "//source/server/config/http:ip_tagging_lib",
         "//source/server/config/http:router_lib",
-        "//source/server/config/network:http_connection_manager_lib",
         "//source/server/http:health_check_lib",
         "//test/common/upstream:utility_lib",
         "//test/config:utility_lib",
@@ -365,6 +365,7 @@ envoy_cc_test(
     srcs = ["stats_integration_test.cc"],
     deps = [
         ":integration_lib",
+        "//source/extensions/filters/network/hcm:config",
         "//test/test_common:network_utility_lib",
         "//test/test_common:utility_lib",
         "@envoy_api//envoy/config/bootstrap/v2:bootstrap_cc",
