+++ test/integration/stats_integration_test.cc
@@ -1,6 +1,8 @@
 #include "envoy/config/bootstrap/v2/bootstrap.pb.h"
 #include "envoy/config/metrics/v2/stats.pb.h"
 
+#include "common/config/well_known_names.h"
+
 #include "test/integration/integration.h"
 #include "test/test_common/network_utility.h"
 #include "test/test_common/utility.h"
