+++ test/mocks/http/mocks.h
@@ -14,8 +14,6 @@
 #include "envoy/http/filter.h"
 #include "envoy/ssl/connection.h"
 
-#include "common/http/conn_manager_impl.h"
-
 #include "test/mocks/common.h"
 #include "test/mocks/event/mocks.h"
 #include "test/mocks/request_info/mocks.h"
@@ -28,42 +26,6 @@
 
 namespace Envoy {
 namespace Http {
-
-class MockConnectionManagerConfig : public ConnectionManagerConfig {
-public:
-  MockConnectionManagerConfig();
-  ~MockConnectionManagerConfig();
-
-  // Http::ConnectionManagerConfig
-  ServerConnectionPtr createCodec(Network::Connection& connection, const Buffer::Instance& instance,
-                                  ServerConnectionCallbacks& callbacks) override {
-    return ServerConnectionPtr{createCodec_(connection, instance, callbacks)};
-  }
-
-  MOCK_METHOD0(accessLogs, const std::list<AccessLog::InstanceSharedPtr>&());
-  MOCK_METHOD3(createCodec_, ServerConnection*(Network::Connection&, const Buffer::Instance&,
-                                               ServerConnectionCallbacks&));
-  MOCK_METHOD0(dateProvider, DateProvider&());
-  MOCK_METHOD0(drainTimeout, std::chrono::milliseconds());
-  MOCK_METHOD0(filterFactory, FilterChainFactory&());
-  MOCK_METHOD0(generateRequestId, bool());
-  MOCK_METHOD0(idleTimeout, const absl::optional<std::chrono::milliseconds>&());
-  MOCK_METHOD0(routeConfigProvider, Router::RouteConfigProvider&());
-  MOCK_METHOD0(serverName, const std::string&());
-  MOCK_METHOD0(stats, ConnectionManagerStats&());
-  MOCK_METHOD0(tracingStats, ConnectionManagerTracingStats&());
-  MOCK_METHOD0(useRemoteAddress, bool());
-  MOCK_CONST_METHOD0(xffNumTrustedHops, uint32_t());
-  MOCK_METHOD0(forwardClientCert, Http::ForwardClientCertType());
-  MOCK_CONST_METHOD0(setCurrentClientCertDetails,
-                     const std::vector<Http::ClientCertDetailsType>&());
-  MOCK_METHOD0(localAddress, const Network::Address::Instance&());
-  MOCK_METHOD0(userAgent, const absl::optional<std::string>&());
-  MOCK_METHOD0(tracingConfig, const Http::TracingConnectionManagerConfig*());
-  MOCK_METHOD0(listenerStats, ConnectionManagerListenerStats&());
-  MOCK_CONST_METHOD0(proxy100Continue, bool());
-  MOCK_CONST_METHOD0(http1Settings, const Http::Http1Settings&());
-};
 
 class MockConnectionCallbacks : public virtual ConnectionCallbacks {
 public:
