+++ test/server/BUILD
@@ -133,10 +133,10 @@ envoy_cc_test(
         "//source/common/network:listen_socket_lib",
         "//source/common/network:socket_option_lib",
         "//source/extensions/filters/listener/original_dst:config",
+        "//source/extensions/filters/network/hcm:config",
         "//source/extensions/transport_sockets/raw_buffer:config",
         "//source/extensions/transport_sockets/ssl:config",
         "//source/server:listener_manager_lib",
-        "//source/server/config/network:http_connection_manager_lib",
         "//test/mocks/server:server_mocks",
         "//test/test_common:environment_lib",
         "//test/test_common:threadsafe_singleton_injector_lib",
@@ -157,6 +157,7 @@ envoy_cc_test(
         "//source/extensions/access_loggers/file:config",
         "//source/extensions/filters/http/dynamo:config",
         "//source/extensions/filters/http/ratelimit:config",
+        "//source/extensions/filters/network/hcm:config",
         "//source/extensions/filters/network/redis_proxy:config",
         "//source/extensions/stat_sinks/statsd:config",
         "//source/server:server_lib",
