+++ test/server/connection_handler_test.cc
@@ -1,6 +1,7 @@
 #include "common/common/utility.h"
 #include "common/network/address_impl.h"
 #include "common/network/raw_buffer_socket.h"
+#include "common/network/utility.h"
 #include "common/stats/stats_impl.h"
 
 #include "server/connection_handler_impl.h"
