+++ test/tools/router_check/router.cc
@@ -5,7 +5,8 @@
 #include <string>
 #include <unordered_map>
 
-#include "common/config/rds_json.h"
+#include "common/network/utility.h"
+#include "common/request_info/request_info_impl.h"
 
 #include "test/test_common/printers.h"
